Breaking up client_wrapper.py into smaller files
